rename model files and update the imports
