+++ ui/src/app/components/add-index-form/add-index-form.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core'
 import { FormArray, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms'
-import IConv, { ICreateIndex } from 'src/app/model/Conv'
+import IConv, { ICreateIndex } from 'src/app/model/conv'
 import { DataService } from 'src/app/services/data/data.service'
 import { SidenavService } from 'src/app/services/sidenav/sidenav.service'
