+++ ui/src/app/components/load-session/load-session.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core'
 import { FormControl, FormGroup, Validators } from '@angular/forms'
 import { DataService } from 'src/app/services/data/data.service'
-import ISessionConfig from '../../model/SessionConfig'
+import ISessionConfig from '../../model/session-config'
 import { Router } from '@angular/router'
 import { InputType, StorageKeys } from 'src/app/app.constants'
 import { extractSourceDbName } from 'src/app/utils/utils'
