+++ ui/src/app/components/object-detail/object-detail.component.ts
@@ -1,17 +1,17 @@
 import { Component, EventEmitter, Input, OnInit, Output, SimpleChanges } from '@angular/core'
 import { FormArray, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms'
-import IUpdateTable from '../../model/UpdateTable'
+import IUpdateTable from '../../model/update-table'
 import { DataService } from 'src/app/services/data/data.service'
 import { MatDialog } from '@angular/material/dialog'
 import { InfodialogComponent } from '../infodialog/infodialog.component'
-import IColumnTabData, { IIndexData } from '../../model/EditTable'
+import IColumnTabData, { IIndexData } from '../../model/edit-table'
 import { SnackbarService } from 'src/app/services/snackbar/snackbar.service'
-import IFkTabData from 'src/app/model/FkTabData'
+import IFkTabData from 'src/app/model/fk-tab-data'
 import { ObjectExplorerNodeType, StorageKeys } from 'src/app/app.constants'
-import FlatNode from 'src/app/model/SchemaObjectNode'
+import FlatNode from 'src/app/model/schema-object-node'
 import { Subscription, take } from 'rxjs'
 import { MatTabChangeEvent } from '@angular/material/tabs/tab-group'
-import IConv from 'src/app/model/Conv'
+import IConv from 'src/app/model/conv'
 import { DropIndexDialogComponent } from '../drop-index-dialog/drop-index-dialog.component'
 
 @Component({
