+++ ui/src/app/components/rule/rule.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { RuleComponent } from './rule.component'
 import { MatTableModule } from '@angular/material/table'
 import { MatIconModule } from '@angular/material/icon'
-import IRuleContent from 'src/app/model/Rule'
+import IRuleContent from 'src/app/model/rule'
 import { By } from '@angular/platform-browser'
 import { HttpClientModule } from '@angular/common/http'
 import { DataService } from 'src/app/services/data/data.service'
