+++ ui/src/app/components/summary/summary.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit, Output, SimpleChanges, EventEmitter } from '@angular/core'
 import { DataService } from 'src/app/services/data/data.service'
-import ISummary, { ISummaryRow } from 'src/app/model/Summary'
-import { FlatNode } from 'src/app/model/SchemaObjectNode'
+import ISummary, { ISummaryRow } from 'src/app/model/summary'
+import { FlatNode } from 'src/app/model/schema-object-node'
 import { Observable, of } from 'rxjs'
 import { map, startWith } from 'rxjs/operators'
 import { FormControl } from '@angular/forms'
