+++ ui/src/app/components/workspace/workspace.component.ts
@@ -2,12 +2,12 @@ import { Component, OnDestroy, OnInit } from '@angular/core'
 import { DataService } from 'src/app/services/data/data.service'
 import { ConversionService } from '../../services/conversion/conversion.service'
 import { SidenavService } from 'src/app/services/sidenav/sidenav.service'
-import IConv from '../../model/Conv'
+import IConv from '../../model/conv'
 import { Subscription } from 'rxjs/internal/Subscription'
 import { MatDialog } from '@angular/material/dialog'
-import IFkTabData from 'src/app/model/FkTabData'
-import IColumnTabData, { IIndexData } from '../../model/EditTable'
-import ISchemaObjectNode, { FlatNode } from 'src/app/model/SchemaObjectNode'
+import IFkTabData from 'src/app/model/fk-tab-data'
+import IColumnTabData, { IIndexData } from '../../model/edit-table'
+import ISchemaObjectNode, { FlatNode } from 'src/app/model/schema-object-node'
 import { ObjectExplorerNodeType } from 'src/app/app.constants'
 
 @Component({
