+++ ui/src/app/model/schema-object-node.ts
