+++ ui/src/app/model/session-config.ts
