+++ ui/src/app/model/spanner-config.ts
