+++ ui/src/app/services/data/data.service.ts
@@ -1,17 +1,17 @@
 import { Injectable } from '@angular/core'
 import { FetchService } from '../fetch/fetch.service'
-import IConv, { ICreateIndex, IInterleaveStatus } from '../../model/Conv'
-import IRuleContent from 'src/app/model/Rule'
+import IConv, { ICreateIndex, IInterleaveStatus } from '../../model/conv'
+import IRuleContent from 'src/app/model/rule'
 import { BehaviorSubject, forkJoin, Observable, of } from 'rxjs'
 import { catchError, filter, map, tap } from 'rxjs/operators'
-import IUpdateTable from 'src/app/model/UpdateTable'
-import IDumpConfig from 'src/app/model/DumpConfig'
-import ISessionConfig from '../../model/SessionConfig'
+import IUpdateTable from 'src/app/model/update-table'
+import IDumpConfig from 'src/app/model/dump-config'
+import ISessionConfig from '../../model/session-config'
 import { InputType, StorageKeys } from 'src/app/app.constants'
-import ISession from 'src/app/model/Session'
-import ISpannerConfig from '../../model/SpannerConfig'
+import ISession from 'src/app/model/session'
+import ISpannerConfig from '../../model/spanner-config'
 import { SnackbarService } from '../snackbar/snackbar.service'
-import ISummary from 'src/app/model/Summary'
+import ISummary from 'src/app/model/summary'
 
 @Injectable({
   providedIn: 'root',
